feat(HYDRA-243): display farming entries as collapsed within single position
