+++ src/i18n/locales/en/translations.json
@@ -38,6 +38,7 @@
   "value.usd": "{{ amount, bignumber(numberPrefix: '$'; type: 'dollar') }}",
   "value.percentage": "{{ value, bignumber(numberSuffix: '%'; type: 'percentage') }}",
   "value.APR": "{{ apr, bignumber(numberSuffix: '% APR'; type: 'percentage') }}",
+  "value.APR.range": "{{ from, bignumber(type: 'percentage') }}-{{ to, bignumber(type: 'percentage') }}% APR",
   "value.multiAPR": "Between {{minApr, bignumber(type: 'percentage')}}-{{maxApr, bignumber(type: 'percentage')}}% APR",
   "farms.deposit.title": "Amount to deposit",
   "farms.deposit.balance": "<0>Share token balance:</0> {{ balance, bignumber(fixedPointScale: 12; type: 'token') }}",
