+++ src/sections/pools/farm/modals/positionDetail/PoolFarmPositionDetail.tsx
@@ -10,6 +10,7 @@ import { DepositNftType } from "api/deposits"
 
 export const PoolFarmPositionDetail = (props: {
   pool: PoolBase
+  depositNft?: DepositNftType
   isOpen: boolean
   onClose: () => void
   onSelect: () => void
