+++ src/sections/pools/pool/shares/deposit/PoolSharesDeposit.styled.ts
@@ -3,15 +3,25 @@ import { theme } from "theme"
 
 export const SContainer = styled.div`
   display: grid;
-  grid-template-columns: 1fr 1fr 1fr 2fr;
+  grid-template-columns: auto 1fr;
   grid-column-gap: 8px;
-`
 
-export const SMobContiner = styled.div`
   padding: 20px;
 
   background-color: rgba(${theme.rgbColors.white}, 0.08);
-
   border: 1px solid rgba(${theme.rgbColors.white}, 0.12);
   border-radius: 14px;
+`
+
+export const SIcon = styled.div`
+  height: 22px;
+
+  display: flex;
+  align-items: center;
+`
+
+export const SPositionContainer = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr auto;
+  grid-column-gap: 8px;
 `
