+++ src/sections/pools/pool/shares/deposit/PoolSharesDepositFarm.styled.ts
@@ -14,3 +14,21 @@ export const SContainer = styled.div<{ isHighlighted?: boolean }>`
     isHighlighted &&
     `background-color: rgba(${theme.rgbColors.primary100}, 0.12);`}
 `
+
+export const SAssetContainer = styled.div`
+  position: relative;
+  display: flex;
+
+  > div:not(:first-of-type) {
+    margin-left: -12px;
+  }
+`
+
+export const SAssetIcon = styled.div`
+  border-radius: 9999px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  background: ${theme.colors.backgroundGray900};
+`
